Dash original 1.0.1
This is an spanish version with all the develop i made so far. On the file it says 1.0.0 but is 1.0.1. it has 10 levels so far 2 bosses the second on development, and feel free to play it  and if you can fix issues
if you upload a video or anything on any social media credit me with https://www.youtube.com/@soloxq_/featured or https://www.youtube.com/@Soloxq_g I will probably add more ways... feel free to use 1 or the 2 of the links